fix(desktop): Build the macOS bundle detector only on darwin
The Windows CI job failed on TestTryMacOSApp_ProbesBothApplicationsRoots.
It expected `/Users/alice/Applications/Visual Studio Code.app` and got
`\Users\alice\Applications\Visual Studio Code.app`.

tryMacOSApp builds .app paths, which are always POSIX, but it composed
them with filepath.Join and expanded the tilde through
pathutil.ExpandHome, and both follow the HOST separator. Only
externalapps_darwin.go calls the detector, so the separator never differs
in production. The portable test file carried its tests, so a Windows
runner executed macOS path logic and asserted a POSIX result.

- Move execMacOSApp, macOSAppBases, and tryMacOSApp into
  externalapps_darwin.go, where the build tag removes the foreign host
- Move the six tryMacOSApp and execMacOSApp tests into
  externalapps_darwin_test.go
- Record why the detector cannot go back to the portable file
- Drop filepath.ToSlash from the darwin tests, where it converts nothing,
  and the import that it needed
- Correct two comments: tryMacOSApp tries every name in ONE base before
  the next base, and the passive and idiomatic sentences become active
