+++ desktop/go/externalapps_darwin_test.go
@@ -3,7 +3,6 @@
 package main
 
 import (
-	"path/filepath"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
@@ -63,7 +62,7 @@ func TestDarwinVSCode_PrefersTheBundleOverThePathCommand(t *testing.T) {
 
 	got := findDarwinSpec(t, "vscode").detect(p)
 	require.NotNil(t, got)
-	assert.Equal(t, "/Applications/Visual Studio Code.app", filepath.ToSlash(got.describe))
+	assert.Equal(t, "/Applications/Visual Studio Code.app", got.describe)
 }
 
 // With no bundle anywhere, the PATH command is still better than nothing: the
@@ -90,7 +89,7 @@ func TestDarwinJetBrains_PrefersTheToolboxBundleOverItsScript(t *testing.T) {
 
 	got := findDarwinSpec(t, "goland").detect(p)
 	require.NotNil(t, got)
-	assert.Equal(t, "/Users/alice/Applications/JetBrains Toolbox/GoLand.app", filepath.ToSlash(got.describe))
+	assert.Equal(t, "/Users/alice/Applications/JetBrains Toolbox/GoLand.app", got.describe)
 }
 
 // findDarwinSpec returns the ExternalAppSpec with the given id from the live
@@ -104,4 +103,85 @@ func findDarwinSpec(t *testing.T, id string) ExternalAppSpec {
 	}
 	t.Fatalf("spec %q not found in defaultExternalAppSpecs()", id)
 	return ExternalAppSpec{}
+}
+
+// --- macOS bundle detection and launch ---
+//
+// These tests assert POSIX .app paths, so they belong to the darwin build and
+// cannot move to the portable test file. See the note above execMacOSApp.
+
+func TestTryMacOSApp_ProbesBothApplicationsRoots(t *testing.T) {
+	t.Parallel()
+	p := newFakeProber()
+	p.setHome("/Users/alice")
+	// Only the user-Applications copy exists.
+	p.addPath("/Users/alice/Applications/Visual Studio Code.app")
+
+	got := tryMacOSApp("Visual Studio Code")(p)
+	require.NotNil(t, got)
+	assert.Equal(t, []string{"open", "-a", "/Users/alice/Applications/Visual Studio Code.app", "/repo"},
+		argvOf(got, "/repo"), "`open -a` is the only route that RAISES the application")
+	assert.Equal(t, "/Users/alice/Applications/Visual Studio Code.app", got.describe,
+		"path field carries the RESOLVED bundle, so `open -a` addresses one exact copy")
+}
+
+func TestTryMacOSApp_PrefersSystemApplicationsOverUserCopy(t *testing.T) {
+	t.Parallel()
+	p := newFakeProber()
+	p.setHome("/Users/alice")
+	p.addPath("/Applications/Cursor.app")
+	p.addPath("/Users/alice/Applications/Cursor.app")
+
+	got := tryMacOSApp("Cursor")(p)
+	require.NotNil(t, got)
+	assert.Equal(t, "/Applications/Cursor.app", got.describe)
+}
+
+// JetBrains Toolbox installs one level below ~/Applications. Without that base
+// a Toolbox user falls through to the wrapper script, which starts the IDE
+// without raising it.
+func TestTryMacOSApp_FindsJetBrainsToolboxBundle(t *testing.T) {
+	t.Parallel()
+	p := newFakeProber()
+	p.setHome("/Users/alice")
+	p.addPath("/Users/alice/Applications/JetBrains Toolbox/GoLand.app")
+
+	got := tryMacOSApp("GoLand")(p)
+	require.NotNil(t, got)
+	assert.Equal(t, "/Users/alice/Applications/JetBrains Toolbox/GoLand.app", got.describe)
+}
+
+// One product, two bundle names: the website's download and the Toolbox copy
+// differ, and Zed ships a Preview channel beside the stable one.
+func TestTryMacOSApp_AcceptsAnyOfSeveralBundleNames(t *testing.T) {
+	t.Parallel()
+	p := newFakeProber()
+	p.setHome("/Users/alice")
+	p.addPath("/Applications/Zed Preview.app")
+
+	got := tryMacOSApp("Zed", "Zed Preview")(p)
+	require.NotNil(t, got)
+	assert.Equal(t, "/Applications/Zed Preview.app", got.describe)
+}
+
+// A "~" base with no home directory would otherwise expand to a RELATIVE path
+// and probe whatever the working directory holds.
+func TestTryMacOSApp_SkipsRelativeCandidateWhenHomeIsEmpty(t *testing.T) {
+	t.Parallel()
+	p := newFakeProber()
+	p.setHome("")
+	p.addPath("Applications/Cursor.app")
+
+	assert.Nil(t, tryMacOSApp("Cursor")(p))
+}
+
+// Only `open -a` activates the target on macOS. The bundle's own command hands
+// the folder to the running instance and exits, and that leaves its window
+// behind this one, so the menu item looks like it does nothing.
+func TestExecMacOSApp_GoesThroughOpenSoTheAppIsRaised(t *testing.T) {
+	t.Parallel()
+	plan := execMacOSApp("/Applications/Visual Studio Code.app").command("/repo")
+
+	assert.Equal(t, []string{"open", "-a", "/Applications/Visual Studio Code.app", "/repo"}, plan.cmd.Args)
+	assert.True(t, plan.exitMeaningful)
 }
